pyproject.toml
[tool.poetry]
name = "data-lib"
version = "0.1.0"
description = ""
authors = ["mosafdar <mosafdaralii@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "2.0.36"
psycopg2 = "2.9.10"
python-dotenv = "1.0.1"
alembic = "1.14.0"
greenlet = "3.1.1"
mako = "1.3.6"
markupsafe = "3.0.2"
psycopg2-binary = "2.9.10"
typing-extensions = "4.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
